requirements.txt
cryptography~=43.0.1
pip~=24.2
attrs~=24.2.0
pillow~=10.4.0
asyncio~=3.4.3
filelock~=3.16.1
cffi~=1.17.1
pytz~=2024.2
typing_extensions~=4.12.2
Django~=5.1.2
requests~=2.32.3
asgiref~=3.8.1
sqlparse~=0.5.1
PyYAML~=6.0.2
certifi~=2024.8.30
tzlocal~=5.2
urllib3~=2.2.3
wcwidth~=0.2.13
six~=1.16.0
idna~=3.10
charset-normalizer~=3.3.2
pycparser~=2.22
APScheduler~=3.10.4
python-dotenv~=1.0.1
celery ~= 5.4.0
django-allauth ~=65.0.2
djangorestframework-simplejwt ~=5.3.1
djangorestframework ~= 3.15.2
markdown ~= 3.7
django-filter ~= 24.3
